Move Metadata from proto to api
