+++ cdap-app-fabric/src/test/java/co/cask/cdap/internal/app/runtime/artifact/ArtifactRepositoryTest.java
@@ -22,6 +22,7 @@ import co.cask.cdap.api.artifact.ArtifactRange;
 import co.cask.cdap.api.artifact.ArtifactVersion;
 import co.cask.cdap.api.macro.MacroFunction;
 import co.cask.cdap.api.macro.Macros;
+import co.cask.cdap.api.metadata.MetadataScope;
 import co.cask.cdap.api.plugin.Plugin;
 import co.cask.cdap.api.plugin.PluginClass;
 import co.cask.cdap.api.plugin.PluginContext;
@@ -59,7 +60,6 @@ import co.cask.cdap.internal.app.runtime.plugin.TestMacroEvaluator;
 import co.cask.cdap.proto.Id;
 import co.cask.cdap.proto.id.Ids;
 import co.cask.cdap.proto.id.NamespaceId;
-import co.cask.cdap.api.metadata.MetadataScope;
 import com.google.common.base.Charsets;
 import com.google.common.base.Preconditions;
 import com.google.common.collect.ImmutableMap;
