+++ cdap-app-fabric/src/test/java/co/cask/cdap/metadata/SystemMetadataAuditPublishTest.java
@@ -17,6 +17,8 @@
 package co.cask.cdap.metadata;
 
 import co.cask.cdap.AllProgramsApp;
+import co.cask.cdap.api.metadata.Metadata;
+import co.cask.cdap.api.metadata.MetadataScope;
 import co.cask.cdap.common.conf.CConfiguration;
 import co.cask.cdap.common.conf.Constants;
 import co.cask.cdap.common.namespace.NamespaceAdmin;
@@ -31,8 +33,6 @@ import co.cask.cdap.proto.audit.AuditPayload;
 import co.cask.cdap.proto.audit.AuditType;
 import co.cask.cdap.proto.audit.payload.metadata.MetadataPayload;
 import co.cask.cdap.proto.id.NamespaceId;
-import co.cask.cdap.proto.metadata.Metadata;
-import co.cask.cdap.api.metadata.MetadataScope;
 import co.cask.cdap.scheduler.Scheduler;
 import com.google.common.base.Predicate;
 import com.google.common.collect.Iterables;
