+++ cdap-client-tests/src/test/java/co/cask/cdap/client/LineageTestRun.java
@@ -16,6 +16,7 @@
 
 package co.cask.cdap.client;
 
+import co.cask.cdap.api.metadata.MetadataScope;
 import co.cask.cdap.client.app.AllProgramsApp;
 import co.cask.cdap.common.BadRequestException;
 import co.cask.cdap.common.app.RunIds;
@@ -35,7 +36,6 @@ import co.cask.cdap.proto.id.DatasetId;
 import co.cask.cdap.proto.id.NamespaceId;
 import co.cask.cdap.proto.id.ProgramId;
 import co.cask.cdap.proto.id.StreamId;
-import co.cask.cdap.api.metadata.MetadataScope;
 import co.cask.cdap.proto.metadata.lineage.CollapseType;
 import co.cask.cdap.proto.metadata.lineage.LineageRecord;
 import co.cask.cdap.test.SlowTests;
