+++ cdap-common/src/main/java/co/cask/cdap/common/metadata/MetadataRecord.java
@@ -16,8 +16,8 @@
 
 package co.cask.cdap.common.metadata;
 
-import co.cask.cdap.proto.id.NamespacedEntityId;
 import co.cask.cdap.api.metadata.MetadataScope;
+import co.cask.cdap.proto.id.NamespacedEntityId;
 
 import java.util.Collections;
 import java.util.Map;
