+++ cdap-data-fabric/src/main/java/co/cask/cdap/data2/metadata/system/SystemMetadataWriter.java
@@ -16,8 +16,8 @@
 
 package co.cask.cdap.data2.metadata.system;
 
-import co.cask.cdap.proto.Id;
 import co.cask.cdap.api.metadata.MetadataScope;
+import co.cask.cdap.proto.Id;
 
 /**
  * Used write {@link MetadataScope#SYSTEM} metadata for an {@link Id.NamespacedId entity}.
