+++ cdap-proto/src/main/java/co/cask/cdap/proto/audit/payload/metadata/MetadataPayload.java
@@ -17,9 +17,9 @@
 package co.cask.cdap.proto.audit.payload.metadata;
 
 import co.cask.cdap.api.annotation.Beta;
+import co.cask.cdap.api.metadata.Metadata;
 import co.cask.cdap.api.metadata.MetadataScope;
 import co.cask.cdap.proto.audit.AuditPayload;
-import co.cask.cdap.proto.metadata.Metadata;
 
 import java.util.Map;
 import java.util.Objects;
